Expanded worldgen limits

+++ src/common/noise.c
@@ -27,35 +27,35 @@ static int noise2(int t, int x, int y) {
     return hash[t][xindex];
 }
 
-static float lin_inter(float x, float y, float s) {
+static double lin_inter(double x, double y, double s) {
     return x + s * (y - x);
 }
 
-static float smooth_inter(float x, float y, float s) {
+static double smooth_inter(double x, double y, double s) {
     return lin_inter(x, y, s * s * (3 - 2 * s));
 }
 
-float noise2d(int tbl, float x, float y) {
+double noise2d(int tbl, double x, double y) {
     const int x_int = floor(x);
     const int y_int = floor(y);
-    const float x_frac = x - x_int;
-    const float y_frac = y - y_int;
+    const double x_frac = x - x_int;
+    const double y_frac = y - y_int;
     const int s = noise2(tbl, x_int, y_int);
     const int t = noise2(tbl, x_int + 1, y_int);
     const int u = noise2(tbl, x_int, y_int + 1);
     const int v = noise2(tbl, x_int + 1, y_int + 1);
-    const float low = smooth_inter(s, t, x_frac);
-    const float high = smooth_inter(u, v, x_frac);
-    const float result = smooth_inter(low, high, y_frac);
+    const double low = smooth_inter(s, t, x_frac);
+    const double high = smooth_inter(u, v, x_frac);
+    const double result = smooth_inter(low, high, y_frac);
     return result;
 }
 
-float perlin2d(int t, float x, float y, float freq, int depth) {
-    float xa = x * freq;
-    float ya = y * freq;
-    float amp = 1.0;
-    float fin = 0;
-    float div = 0.0;
+double perlin2d(int t, double x, double y, double freq, int depth) {
+    double xa = x * freq;
+    double ya = y * freq;
+    double amp = 1.0;
+    double fin = 0;
+    double div = 0.0;
     for (int i = 0; i < depth; i++) {
         div += 256 * amp;
         fin += noise2d(t, xa, ya) * amp;
@@ -66,9 +66,9 @@ float perlin2d(int t, float x, float y, float freq, int depth) {
     return fin/div;
 }
 
-float mperlin2d(int t, float x, float y, float freq, int depth, int samples) {
-    float s = 0.0;
-    float div = (1 + samples * 2) + samples * samples * 2;
+double mperlin2d(int t, double x, double y, double freq, int depth, int samples) {
+    double s = 0.0;
+    double div = (1 + samples * 2) + samples * samples * 2;
     //printf("BEGIN: [%lf]\n", div);
     for (int i = -samples; i <= samples; ++i) {
         int s2 = samples - abs(i);
